Move the macro into lib.rs
