+++ yewprint-doc/src/buttons/mod.rs
@@ -5,68 +5,6 @@ use example::*;
 use yew::prelude::*;
 use yewprint::{Switch, Text, H1, H5};
 
-macro_rules! build_source_code_component {
-    ($constant_name:ident, $url:expr, $test_name:ident) => {
-        const $constant_name: &'static str = $url;
-
-        pub struct SourceCodeUrl {
-            props: SourceCodeUrlProps,
-        }
-
-        #[derive(Clone, PartialEq, Properties)]
-        pub struct SourceCodeUrlProps {
-            #[prop_or_default]
-            pub url: &'static str,
-        }
-
-        impl Component for SourceCodeUrl {
-            type Message = ();
-            type Properties = SourceCodeUrlProps;
-
-            fn create(props: Self::Properties, _link: ComponentLink<Self>) -> Self {
-                Self { props }
-            }
-
-            fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-                true
-            }
-
-            fn change(&mut self, props: Self::Properties) -> ShouldRender {
-                if self.props != props {
-                    self.props = props;
-                    true
-                } else {
-                    false
-                }
-            }
-
-            fn view(&self) -> Html {
-                html! {
-                    <a
-                        class=classes!("bp3-text-muted")
-                        href=self.props.url
-                        target="_blank"
-                    >
-                        <Text>{"Go to the source code"}</Text>
-                    </a>
-                }
-            }
-        }
-
-        #[cfg(test)]
-        mod tests {
-            use super::*;
-
-            #[test]
-            fn $test_name() {
-                let get_url = reqwest::blocking::get($constant_name).unwrap();
-
-                assert!(get_url.status().is_success())
-            }
-        }
-    };
-}
-
 pub struct ButtonDoc {
     callback: Callback<ExampleProps>,
     state: ExampleProps,
